Optimise the state toggling logic
Previously we were issuing three change requests for every review action,
e.g. running `toggle-approved` would:

1.  look at whether a photo was in the 'approved' album, and add/remove
    as appropriate
2.  remove a photo from the 'rejected' album
3.  remove a photo from the 'needs action' album

But in practice actions (2) and (3) are usually redundant; most of what
I'm reviewing is unreviewed and isn't in any of the three albums.
Since we're already getting a complete list of albums that contain this
photo, we can check which albums need to change and only issue the
appropriate change requests.

I haven't profiled extensively, but in informal testing this does seem
to make the action maybe ~25% faster?