Simplify how the FocusedImage is determined
Putting an @ObservableObject on the PhotoReviewer View was a mistake;
it causes the view to reload whenever it changes, and it was in the same
view as the thumbnail list.  Moving it down into the FocusedImage view
means we're doing less reloading whenever the focused image changes.

This has a couple of knock-on effects:

*   Paging back/forth isn't painfully slow when you're navigated deep into
    the photo library -- it's still a little laggy, but not quite as bad
    as it was.

*   The new focused image updates immediately, rather than waiting for
    the full-sized version to load over the network.  There are some
    mitigations to ensure this still looks good.