Fix a bug where the thumbnail list got out of sync
When you deleted photos, the `assetIdentifiers` list wasn't being
regenerated properly -- it would only regenerate if assets were being
moved, which isn't the same as deleted.

This meant that when you deleted photos, they'd stay in the thumbnail
list but the main photo would be different.

Adding an extra condition that forces a regeneration fixes this issue.