Fix another bug where the thumbnail list got out of sync
In the previous commit, we added a regeneration when assets were deleted,
but we also need to do it when assets are added.  Note that new assets
being added may not just occur when the user does an explicit action;
it can also occur if an iCloud sync occurs in the background and fetches
new photos from the cloud.